refactor: avoid song model namespace conflict
