+++ KaraIOke/Views/Templates/SongTemplate.xaml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="UTF-8"?>
 <ContentView xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:models="clr-namespace:KaraIOke.Models.Song"
+             xmlns:models="clr-namespace:KaraIOke.Models"
              xmlns:views="clr-namespace:KaraIOke.Views"
              xmlns:viewmodels="clr-namespace:KaraIOke.ViewModels"
              x:Class="KaraIOke.Views.Templates.SongTemplate"
